docs: Add basic build docs
