+++ README.md
@@ -1,8 +1,27 @@
 # consumat.io-backend
+
 Backend for https://github.com/alphahorizonio/consumat.io.
+
 ## Installation
+
 🚧 This project is a work-in-progress! Instructions will be added as soon as it is usable. 🚧
-## License 
-consumat.io-backend (c) 2021 Jakob Waibel and contributors 
+
+## Contributing
+
+To contribute, please use the [GitHub flow](https://guides.github.com/introduction/flow/).
+
+To build and start a development version locally, run the following:
+
+```shell
+$ git clone https://github.com/alphahorizonio/consumat.io-backend.git
+$ pip install -r requirements.txt
+$ uvicorn consumatio.external.api:app
+```
+
+The backend's GraphQL Playground should now be available on [http://localhost:8000/](http://localhost:8000/).
+
+## License
+
+consumat.io-backend (c) 2021 Jakob Waibel and contributors
 
 SPDX-License-Identifier: AGPL-3.0
