+++ consumatio/external/api.py
@@ -1,116 +1,113 @@
-from graphene import ObjectType, String, Schema, Decimal, Int, List, Field, Boolean, Float
-from flask import Flask, request
-import json
+from ariadne import ObjectType, QueryType, gql, make_executable_schema
+from ariadne.asgi import GraphQL
 from consumatio.external.tmdb import Tmdb
-from consumatio.usecases.movie_details import *
+from consumatio.usecases.movie_details import * 
 from consumatio.usecases.tv_details import *
 from consumatio.usecases.season_details import *
 from consumatio.usecases.episode_details import *
-import dataclasses
-from flask_cors import CORS
 
-app = Flask(__name__)
-CORS(app)
+type_defs = gql("""
+    type Query {
+        movie(code: Int, country: String): Movie!
+        tv(code: Int, country: String): TV!
+        season(code: Int, seasonNumber: Int): Season!
+        episode(code: Int, seasonNumber: Int, episodeNumber: Int): Episode!
+    }
 
-class Movie(ObjectType):
-    code = Int()
-    title = String()
-    genres = List(String)
-    overview = String()
-    popularity = Float()
-    vote_average = Float()
-    release_date = String()
-    runtime = Int()
-    status = String()
-    backdrops = List(String)
-    posters = List(String)
-    providers = List(String)
-    watch_status = String()
-    rating = Float()
-    favorite = Boolean()
+    type Movie {
+        code: Int
+        title: String
+        genres: [String]
+        overview: String
+        popularity: Float
+        voteAverage: Float
+        releaseDate: String
+        runtime: Int
+        status: String
+        backdrops: String
+        posters: String
+        providers: [String]
+        watchStatus: String
+        rating: Float
+        favorite: Boolean
+    }
 
-    @classmethod 
-    def from_dict(self, dict):
-        return self(**dict)
+    type TV {
+        code: Int
+        name: String
+        genres: [String]
+        overview: String
+        popularity: Float
+        voteAverage: Float
+        firstAirDate: String
+        lastAirDate: String
+        status: String
+        backdrops: String
+        posters: String
+        providers: [String]
+        watchStatus: String
+        rating: Float
+        favorite: Boolean
+    }
 
-class TV(ObjectType):
-    code = Int()
-    name = String()
-    genres = List(String)
-    overview = String()
-    popularity = Float()
-    vote_average = Float()
-    first_air_date = String()
-    last_air_date = String()
-    status = String()
-    backdrops = List(String)
-    posters = List(String)
-    providers = List(String)
-    watch_status = String()
-    rating = Float()
-    favorite = Boolean()
+    type Season {
+        code: Int
+        tvCode: Int
+        seasonNumber: Int
+        name: String
+        overview: String
+        posters: String
+        watchStatus: String
+        rating: Float
+        favorite: Boolean
+    }
 
-    @classmethod 
-    def from_dict(self, dict):
-        return self(**dict)
+    type Episode {
+        code: Int
+        name: String
+        episodeNumber: Int
+        seasonNumber: Int
+        overview: String
+        airDate: String
+        voteAverage: Float 
+        stills: String
+        watchStatus: String
+        rating: Float
+        favorite: Boolean
+    }
+""")
 
-class Season(ObjectType):
-    code = Int()
-    tv_code = Int()
-    season_number = Int()
-    name = String()
-    overview = String()
-    posters = List(String)
-    watch_status = String()
-    rating = Float()
-    favorite = Boolean()
 
-    @classmethod 
-    def from_dict(self, dict):
-        return self(**dict)
+query = QueryType()
 
-class Episode(ObjectType):
-    code = Int()
-    name = String()
-    episode_number = Int()
-    season_number = Int()
-    overview = String()
-    air_date = String()
-    vote_average = Float()
-    stills = List(String)
-    watch_status = String()
-    rating = Float()
-    favorite = Boolean()
+@query.field("movie")
+def resolve_movie(*_, code, country):
+    tmdb = Tmdb()
+    return  movie_details(tmdb, code, country)
 
-    @classmethod 
-    def from_dict(self, dict):
-        return self(**dict)
+movie = ObjectType("Movie")
 
-class Query(ObjectType):
-    movie_details = Field(Movie, code=Int(), country=String())
-    tv_details = Field(TV, code=Int(), country=String())
-    season_details = Field(Season, code=Int(), season_number=Int())
-    episode_details = Field(Episode, code=Int(), season_number=Int(), episode_number=Int())
+@query.field("tv")
+def resolve_tv(*_, code, country):
+    tmdb = Tmdb()
+    return tv_details(tmdb, code, country)
 
-    def resolve_movie_details(root, info, code, country):
-        tmdb = Tmdb()
-        return Movie.from_dict(movie_details(tmdb, code, country))
+tv = ObjectType("TV")
 
-    def resolve_tv_details(root, info, code, country):
-        tmdb = Tmdb()
-        return TV.from_dict(tv_details(tmdb, code, country))
+@query.field("season")
+def resolve_season(*_, code, seasonNumber):
+    tmdb = Tmdb()
+    return season_details(tmdb, code, seasonNumber)
 
-    def resolve_season_details(root, info, code, season_number):
-        tmdb = Tmdb()
-        return Season.from_dict(season_details(tmdb, code, season_number))
+season = ObjectType("Season")
 
-    def resolve_episode_details(root, info, code, season_number, episode_number):
-        tmdb = Tmdb()
-        return Episode.from_dict(episode_details(tmdb, code, season_number, episode_number))
+@query.field("episode")
+def resolve_episode(*_, code, seasonNumber, episodeNumber):
+    tmdb = Tmdb()
+    return episode_details(tmdb, code, seasonNumber, episodeNumber)
 
-schema = Schema(query=Query)
+episode = ObjectType("Episode")
 
-@app.route('/graphql', methods=['POST', 'GET'])
-def graphql():
-    data = json.loads(request.data)
-    return json.dumps(schema.execute(data['query']).data)
+schema = make_executable_schema(type_defs, query, movie, tv, season)
+
+app = GraphQL(schema, debug=True)
