--- consumatio/external/api_ariadne.py
@@ -1,113 +1,0 @@
-from ariadne import ObjectType, QueryType, gql, make_executable_schema
-from ariadne.asgi import GraphQL
-from consumatio.external.tmdb import Tmdb
-from consumatio.usecases.movie_details import * 
-from consumatio.usecases.tv_details import *
-from consumatio.usecases.season_details import *
-from consumatio.usecases.episode_details import *
-
-type_defs = gql("""
-    type Query {
-        movie(code: Int, country: String): Movie!
-        tv(code: Int, country: String): TV!
-        season(code: Int, seasonNumber: Int): Season!
-        episode(code: Int, seasonNumber: Int, episodeNumber: Int): Episode!
-    }
-
-    type Movie {
-        code: Int
-        title: String
-        genres: [String]
-        overview: String
-        popularity: Float
-        voteAverage: Float
-        releaseDate: String
-        runtime: Int
-        status: String
-        backdrops: String
-        posters: String
-        providers: [String]
-        watchStatus: String
-        rating: Float
-        favorite: Boolean
-    }
-
-    type TV {
-        code: Int
-        name: String
-        genres: [String]
-        overview: String
-        popularity: Float
-        voteAverage: Float
-        firstAirDate: String
-        lastAirDate: String
-        status: String
-        backdrops: String
-        posters: String
-        providers: [String]
-        watchStatus: String
-        rating: Float
-        favorite: Boolean
-    }
-
-    type Season {
-        code: Int
-        tvCode: Int
-        seasonNumber: Int
-        name: String
-        overview: String
-        posters: String
-        watchStatus: String
-        rating: Float
-        favorite: Boolean
-    }
-
-    type Episode {
-        code: Int
-        name: String
-        episodeNumber: Int
-        seasonNumber: Int
-        overview: String
-        airDate: String
-        voteAverage: Float 
-        stills: String
-        watchStatus: String
-        rating: Float
-        favorite: Boolean
-    }
-""")
-
-
-query = QueryType()
-
-@query.field("movie")
-def resolve_movie(*_, code, country):
-    tmdb = Tmdb()
-    return  movie_details(tmdb, code, country)
-
-movie = ObjectType("Movie")
-
-@query.field("tv")
-def resolve_tv(*_, code, country):
-    tmdb = Tmdb()
-    return tv_details(tmdb, code, country)
-
-tv = ObjectType("TV")
-
-@query.field("season")
-def resolve_season(*_, code, seasonNumber):
-    tmdb = Tmdb()
-    return season_details(tmdb, code, seasonNumber)
-
-season = ObjectType("Season")
-
-@query.field("episode")
-def resolve_episode(*_, code, seasonNumber, episodeNumber):
-    tmdb = Tmdb()
-    return episode_details(tmdb, code, seasonNumber, episodeNumber)
-
-episode = ObjectType("Episode")
-
-schema = make_executable_schema(type_defs, query, movie, tv, season)
-
-app = GraphQL(schema, debug=True)
