+++ requirements.txt
@@ -1,0 +1,2 @@
+ariadne==0.13.0
+uvicorn==0.13.4
